requirements.txt
certifi==2022.6.15
chardet==5.0.0
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
SQLAlchemy==1.4.40
urllib3==1.26.11
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
python-dotenv==0.20.0
gunicorn==20.1.0
psycopg2==2.9.3